pyproject.toml
[project]
name = "autovoicecollation"
version = "0.1.0"
description = "AutoVoiceCollation 是一个功能强大的自动语音识别（ASR）和文本处理工具，旨在帮助用户快速整理和润色语音转录文本。"
requires-python = "==3.11"
dependencies = [
    "cerebras>=0.0.1",
    "dotenv>=0.9.9",
    "fastapi>=0.119.1",
    "funasr>=1.2.7",
    "gradio>=5.49.1",
    "jieba>=0.42.1",
    "openai>=2.6.0",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "protobuf>=6.33.0",
    "pydantic>=2.11.9",
    "pyinstaller>=6.16.0",
    "python-dotenv>=1.1.1",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
    "soundfile>=0.13.1",
    "transformers>=4.56.2",
    "uvicorn>=0.38.0",
    "yt-dlp>=2025.10.14",
    "google-genai>=1.46.0",
    # The following dependencies are commented out to avoid potential installation issues.
    # plz handle torch installation separately
    #    "torch>=2.8.0",
    #    "torchaudio>=2.8.0"
]
